requirements.txt
ffmpeg_python==0.2.0
exif==1.6.0
pymediainfo==6.1.0